solve fetch issue in customers and payments route
