+++ src/app/(root)/customers/page.tsx
@@ -1,0 +1,10 @@
+import CustomersTable from "@/components/CustomersTable";
+
+export default function CustomersPage() {
+  return (
+    <div>
+      <h1 className="text-xl font-bold mb-4">Customers</h1>
+      <CustomersTable />
+    </div>
+  );
+}
